+++ src/controllers/AbstractController.php
@@ -10,8 +10,10 @@ namespace flipbox\craft\restful\controllers;
 
 use craft\helpers\ArrayHelper;
 use flipbox\craft\rest\Controller;
+use flipbox\craft\restful\filters\ResponseLogger;
 use flipbox\craft\restful\Restful;
 use yii\filters\AccessControl;
+use yii\log\Logger;
 
 /**
  * @author Flipbox Factory <hello@flipboxfactory.com>
@@ -35,6 +37,22 @@ abstract class AbstractController extends Controller
                     'except' => [
                         'options',
                         'head'
+                    ]
+                ],
+                'log' => [
+                    'class' => ResponseLogger::class,
+                    'actions' => [
+                        '*' => [
+                            Logger::LEVEL_INFO => [
+                                [200, '>='], // All 4xx status codes
+                                [300, '<']
+                            ],
+                            Logger::LEVEL_ERROR, [
+                                [400, '>='], // All 4xx status codes
+                                [600, '<'], // All 5xx status codes
+                                [404, '!=']
+                            ]
+                        ]
                     ]
                 ]
             ]